feat: JDS Blog scrape
